Created service for e-Cert generation
