+++ sources/packages/backend/libs/integrations/src/constants/system-configurations-constants.ts
@@ -1,0 +1,6 @@
+/**
+ * Amount of days before a disbursement schedule date is due that
+ * it could be included in the e-Cert files to be sent to the
+ * federal government.
+ */
+export const DISBURSEMENT_FILE_GENERATION_ANTICIPATION_DAYS = 5;
