+++ sources/packages/backend/libs/integrations/src/services/disbursement-schedule-service/e-cert-generation.models.ts
@@ -1,0 +1,5 @@
+import { DisbursementSchedule } from "@sims/sims-db";
+
+export interface ECertDisbursementSchedule extends DisbursementSchedule {
+  stopFullTimeBCFunding: boolean;
+}
