+++ sources/packages/backend/libs/integrations/src/services/index.ts
@@ -2,3 +2,5 @@ export * from "./ssh/ssh.service";
 export * from "./ssh/sftp-integration-base";
 export * from "./student-assessment/student-assessment.service";
 export * from "./cra-income-verification/cra-income-verification.service";
+export * from "./disbursement-schedule-service/e-cert-generation-service";
+export * from "./disbursement-schedule-service/e-cert-generation.models";
