+++ sources/packages/backend/libs/sims-db/src/entities/disbursement-schedule.model.ts
@@ -71,7 +71,7 @@ export class DisbursementSchedule extends RecordDataModel {
     (disbursementValue) => disbursementValue.disbursementSchedule,
     {
       eager: false,
-      cascade: true,
+      cascade: ["insert", "update"],
       onDelete: "CASCADE",
       nullable: true,
     },
